backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt==5.3.1
Faker
psycopg2
python-dotenv
drf-yasg
google-auth
django-googledrive-storage
google-api-python-client
django-environ